src/app/requirements.txt
firebase_admin==6.1.0
pydantic==1.10.7
google-cloud-pubsub==2.17.1